add some logic
