+++ csrc/gpu/mla_attn/batch_mla_with_paged_kv_cache.cu
@@ -34,13 +34,6 @@ using namespace cute;
 using namespace mla_attn;
 using namespace std;
 
-inline uint32_t get_max_partition_size(int bsz) {
-    static const char* max_partition_size_env = std::getenv("FLAGS_cascade_attention_max_partition_size");
-    static const uint32_t max_partition_size =
-            max_partition_size_env == nullptr ? 0 : std::stoul(std::string(max_partition_size_env));
-    return (max_partition_size != 0 ? max_partition_size : (bsz == 1 ? 128 : 512));
-}
-
 template <typename T>
 struct cascade_type_traits {
   using type = T;
